Move plate to being a handler
Idea: ensure appropriate batch dimensions before passing forward.
